Carve NetworkFirstPersonController into per-responsibility partials
Splits the 1321-line player controller into a partial class spread
across four files. Zero call-site changes; all members stay private/
public on the same class, just physically located by responsibility.

New partial files (all under Assets/Scripts/Player/):

- NetworkFirstPersonController.Diagnostics.cs (337 lines): the
  optional F9/F10 physics CSV logger. Owns enablePhysicsDiagnostics
  + 4 sibling SerializeFields, the diagnostic state cache (record
  flag, log path, collider buffers, StringBuilder), and all of the
  HandleDiagnosticHotkeys / Start/Stop/Sample/WritePhysicsDiagnostic
  methods plus the CSV format helpers. The "last*" movement-state
  fields stay in the main file because movement writes them.

- NetworkFirstPersonController.PlayerCarrying.cs (155 lines): the
  Lethal-Company-style "carry another player" subsystem. Owns
  carryPlayerSpeedMultiplier, IsBeingCarried + CarriedByClientId
  NetworkVariables, _heldPlayer, the HeldPlayer/HasHeldPlayer/
  CarryPlayerSpeedMultiplier properties, ResolveHeldPlayer, and all
  the RequestPickupByPlayer/MoveHeldPlayer/RequestDropHeldPlayer/
  ServerDropHeldPlayer/ServerMoveAsCarried/ServerReleaseFromCarry
  RPCs and their corresponding ClientRpcs.

- NetworkFirstPersonController.Health.cs (201 lines): health,
  damage, death, revive, and the post-death spectate camera. Owns
  maxHealth, the _health NetworkVariable, the dead/spectating
  state, the IsDead/HealthPercent/IsDeadLocally/SpectatorTargetLabel
  properties, ServerTakeDamage / ServerRevive and the corresponding
  Die/Revive/SpawnBloodSplatter ClientRpcs, plus HandleDeathCamera
  and CycleSpectateTarget.

Main NetworkFirstPersonController.cs goes from 1321 to 773 lines.
Movement, look, crouch, stamina, knockback/stun, surface alignment,
teleport, lifecycle, and name sync stay in the main file - those
form a cohesive "what the player physically does each frame" unit.

Verified with dotnet build on the Runtime project: 0 errors.
