+++ Space Game/Assets/Scripts/Player/NetworkFirstPersonController.PlayerCarrying.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: ef7a2aff542b4d0b8fafc6901df2d494
